Former-commit-id: 9b39c6d94a0e18ecf2c03d471ab95211791e4ecf

+++ Control/StoreGate/src/SGHiveMgrSvc.cxx
@@ -133,7 +133,7 @@ size_t HiveMgrSvc::getPartitionNumber(int evtNumber) const {
   return std::string::npos;
 }
 
-unsigned int HiveMgrSvc::freeSlots() {
+size_t HiveMgrSvc::freeSlots() {
   return m_freeSlots;
 }
